requirements.txt
faker 
pydantic 
fastapi 
uvicorn 
requests
numpy 
faiss-cpu